requirements.txt
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
colorama==0.3.7
cycler==0.10.0
Cython==0.24.1
dask==0.11.0
decorator==4.0.10
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
matplotlib==1.5.3
networkx==1.11
numpy==1.11.1+mkl
pandas==0.18.1
pathlib2==2.1.0
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
Pygments==2.1.3
pyparsing==2.1.9
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.1
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
toolz==0.8.0
traitlets==4.3.0
wcwidth==0.1.7
win-unicode-console==0.5
